# docs/swagger.yaml
definitions:
  problemdetail.ProblemDetail:
    properties:
      context: {}
      detail:
        type: string
      title:
        type: string
      type:
        type: string
    type: object
info:
  contact: {}
paths:
  /api/v1/projects:
    get:
      consumes:
      - application/json
      description: List Projects available on the platform
      operationId: ListProject
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/problemdetail.ProblemDetail'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/problemdetail.ProblemDetail'
      summary: List Projects available on the platform
      tags:
      - projects
    post:
      consumes:
      - application/json
      description: Create a Project
      operationId: CreateProject
      parameters:
      - description: Project Creation data
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/problemdetail.ProblemDetail'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/problemdetail.ProblemDetail'
      summary: Create a Project
      tags:
      - projects
  /api/v1/projects/{projectID}:
    delete:
      consumes:
      - application/json
      description: Delete a Project
      operationId: DeleteProject
      parameters:
      - description: Project ID
        in: path
        name: projectID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/problemdetail.ProblemDetail'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/problemdetail.ProblemDetail'
      summary: Delete a Project
      tags:
      - projects
    get:
      consumes:
      - application/json
      description: Get a single Project
      operationId: GetOneProject
      parameters:
      - description: Project ID
        in: path
        name: projectID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/problemdetail.ProblemDetail'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/problemdetail.ProblemDetail'
      summary: Get a single Project
      tags:
      - projects
    post:
      consumes:
      - application/json
      description: Update a Project
      operationId: UpdateProject
      parameters:
      - description: Project ID
        in: path
        name: projectID
        required: true
        type: string
      - description: Project Update data
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/problemdetail.ProblemDetail'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/problemdetail.ProblemDetail'
      summary: Update a Project
      tags:
      - projects
  /api/v1/projects/import:
    post:
      consumes:
      - application/json
      description: Import Projects from some source
      operationId: ImportProject
      parameters:
      - description: Import Specification
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/problemdetail.ProblemDetail'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/problemdetail.ProblemDetail'
      summary: Import Projects from some source
      tags:
      - projects
  /api/v1/projects/query:
    get:
      consumes:
      - application/json
      description: Search for a Project
      operationId: SearchProjects
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/problemdetail.ProblemDetail'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/problemdetail.ProblemDetail'
      summary: Search for a Project
      tags:
      - projects
  /api/v1/test-cases:
    get:
      consumes:
      - application/json
      description: List Test Cases
      operationId: ListTestCases
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/problemdetail.ProblemDetail'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/problemdetail.ProblemDetail'
      summary: List Test Cases
      tags:
      - test-cases
    post:
      consumes:
      - application/json
      description: Create a new Test Case
      operationId: CreateTestCase
      parameters:
      - description: Create Test Case data
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/problemdetail.ProblemDetail'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/problemdetail.ProblemDetail'
      summary: Create a new Test Case
      tags:
      - test-cases
  /api/v1/test-cases/{testCaseID}:
    delete:
      consumes:
      - application/json
      description: Delete a test case
      operationId: DeleteTestCase
      parameters:
      - description: Test Case ID
        in: path
        name: testCaseID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/problemdetail.ProblemDetail'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/problemdetail.ProblemDetail'
      summary: Delete a test case
      tags:
      - test-cases
    get:
      consumes:
      - application/json
      description: Get a single Test Case
      operationId: GetOneTestCase
      parameters:
      - description: Test Case ID
        in: path
        name: testCaseID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/problemdetail.ProblemDetail'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/problemdetail.ProblemDetail'
      summary: Get a single Test Case
      tags:
      - test-cases
    post:
      consumes:
      - application/json
      description: Update a Test Case
      operationId: UpdateTestCase
      parameters:
      - description: Test Case ID
        in: path
        name: testCaseID
        required: true
        type: string
      - description: Test Case update data
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/problemdetail.ProblemDetail'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/problemdetail.ProblemDetail'
      summary: Update a Test Case
      tags:
      - test-cases
  /api/v1/test-cases/query:
    get:
      consumes:
      - application/json
      description: Search for Test Cases
      operationId: SearchTestCases
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/problemdetail.ProblemDetail'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/problemdetail.ProblemDetail'
      summary: Search for Test Cases
      tags:
      - test-cases
...
  /api/v1/testers:
    get:
      consumes:
      - application/json
      description: List all Testers
      operationId: ListTesters
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/problemdetail.ProblemDetail'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/problemdetail.ProblemDetail'
      summary: List all Testers
      tags:
      - testers
  /api/v1/testers.query:
    get:
      consumes:
      - application/json
      description: Search all Testers
      operationId: SearchTesters
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/problemdetail.ProblemDetail'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/problemdetail.ProblemDetail'
      summary: Search all Testers
      tags:
      - testers
  /api/v1/testers/{testerID}:
    get:
      consumes:
      - application/json
      description: Get a Tester by ID
      operationId: GetOneTester
      parameters:
      - description: Tester ID
        in: path
        name: testerID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/problemdetail.ProblemDetail'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/problemdetail.ProblemDetail'
      summary: Get a Tester by ID
      tags:
      - testers
  /api/v1/testers/invite/{email}:
    post:
      consumes:
      - application/json
      description: Invite a tester by Email
      operationId: InviteTester
      parameters:
      - description: Email of tester
        in: path
        name: email
        required: true
        type: string
      - description: Invite data
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/problemdetail.ProblemDetail'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/problemdetail.ProblemDetail'
      summary: Invite a tester by Email
      tags:
      - testers
...
swagger: "2.0"